Refactor and fix log replication
Fix multiple issues after correcting understanding of "index" of a log
entry. Specifically, the "index" of a log entry should be the global
index of the entry in the log, as opposed to "local" index with in the
current term. In addition, log index should not be represented using the
index into a slice, because the slice containing LogEntry can only be
viewed as a window of the entire log through out the cluster's lifetime.
I haven't dug into log compaction yet, but I suppose in the case of log
compaction, the slice may be truncated from the left. Thus the index of
each entry must be stored within the entry. These corrections bring to
the following changes:

- Move log to its own type with methods, log entries should not be
  touched from outside these methods.
- Store log index in each entry and send index in requests based on
  global index.
- NextIndex and MatchIndex should be based on global index.